// src/bare_frontend/src/components/Text/Text.types.ts
export enum FontWeights {
  Light = '300',
  Regular = '400',
  Medium = '500',
  SemiBold = '600',
  Bold = '700',
}

export enum FontFamily {
  HammersmithOne = 'Hammersmith One',
  IntegralCF = 'Integral CF',
}

export type TextVariants = 'h1' | 'h2' | 'body' | 'subtitle1' | 'subtitle2' | 'button' | 'label';

export type TextColors = 'black' | 'white' | 'damage' | 'accent';